refactor(gateway): move DNS otel helpers into otel module
Move the DNS lookup attribute helpers out of the event loop and into a
dedicated `otel` module, mirroring how `connlib`'s `otel` module is
organised. Metric construction now also goes through `otel::metrics`.
